fix rounding

+++ mfp/macros/macros.go
@@ -49,17 +49,17 @@ func newMacrosMessage(d *mfp.Diary) (string, error) {
 	if err != nil {
 		return "", err
 	}
-	table.Append([]string{"Carbs", d.Total.Carbs, fmt.Sprintf("%d%%", m.carbs)})
-	table.Append([]string{"Protein", d.Total.Protein, fmt.Sprintf("%d%%", m.protein)})
-	table.Append([]string{"Fat", d.Total.Fat, fmt.Sprintf("%d%%", m.fat)})
+	table.Append([]string{"Carbs", d.Total.Carbs, fmt.Sprintf("%.2f%%", m.carbs)})
+	table.Append([]string{"Protein", d.Total.Protein, fmt.Sprintf("%.2f%%", m.protein)})
+	table.Append([]string{"Fat", d.Total.Fat, fmt.Sprintf("%.2f%%", m.fat)})
 	table.Render()
 	return "```" + buffer.String() + "```", nil
 }
 
 type macroPercentages struct {
-	carbs   int
-	protein int
-	fat     int
+	carbs   float32
+	protein float32
+	fat     float32
 }
 
 func newMacroPercentages(d *mfp.Diary) (macroPercentages, error) {
@@ -70,11 +70,11 @@ func newMacroPercentages(d *mfp.Diary) (macroPercentages, error) {
 	if cErr != nil || pErr != nil || fErr != nil {
 		return m, errors.New("Error parsing macros")
 	}
-	total := carbs + protein + fat
+	total := float32(carbs + protein + fat)
 	m = macroPercentages{
-		carbs:   (100.0 * carbs) / total,
-		protein: (100.0 * protein) / total,
-		fat:     (100.0 * fat) / total,
+		carbs:   float32(100*carbs) / total,
+		protein: float32(100*protein) / total,
+		fat:     float32(100*fat) / total,
 	}
 	return m, nil
 }
